Правки по ревью

+++ src/main/java/com/example/aviasimbir/requestresponse/CreateTicketRequest.java
@@ -3,21 +3,23 @@ package com.example.aviasimbir.requestresponse;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class CreateTicketRequest {
-    private Integer price;
+    private BigDecimal price;
     private Boolean commission;
 
     @JsonCreator
-    public CreateTicketRequest(@JsonProperty("price") Integer price, @JsonProperty("commission") Boolean commission) {
+    public CreateTicketRequest(@JsonProperty("price") BigDecimal price, @JsonProperty("commission") Boolean commission) {
         this.price = price;
         this.commission = commission;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
